third_party/WebKit/LayoutTests/inspector/sass/test-edit-insert-property-expected.txt

Verify that mapping is not valid for misaligned sources.

INITIAL MODELS
===== test-edit-insert-property.css =====
.box1.{
..color:.red;
..margin:.10px;.}

.box2.{
..color:.red;
..margin:.10px;.}

.box3.{
..color:.red;
..margin:.10px;.}

body.{
..font-family:."arial";.}

/*#.sourceMappingURL=test-edit-insert-property.css.map.*/

===== test-edit-insert-property.scss =====
$color:.red;

@for.$i.from.1.through.3.{
.....box#{$i}.{
........color:.$color;
........margin:.10px;
....}
}

body.{
....font-family:."arial";
}


------------------------- TEST: testPrependProperty --------------------------

Edits:
    {1, 2, 1, 7} 'color' => 'border: 1px solid black;color'
===== test-edit-insert-property.css =====
    .box1.{
    ..border:.1px.solid.black;
    ..color:.red;
    ..margin:.10px;.}
    
    .box2.{
    ..border:.1px.solid.black;
    ..color:.red;
    ..margin:.10px;.}
    
    .box3.{
    ..border:.1px.solid.black;
    ..color:.red;
    ..margin:.10px;.}
    
    body.{
    ..font-family:."arial";.}
    
    /*#.sourceMappingURL=test-edit-insert-property.css.map.*/
    
===== test-edit-insert-property.scss =====
    $color:.red;
    
    @for.$i.from.1.through.3.{
    .....box#{$i}.{
    ........border:.1px.solid.black;
    ........color:.$color;
    ........margin:.10px;
    ....}
    }
    
    body.{
    ....font-family:."arial";
    }
    

----------------------- TEST: testInsertMiddleProperty -----------------------

Edits:
    {1, 9, 1, 13} 'red;' => 'red;border: 1px solid black;'
===== test-edit-insert-property.css =====
    .box1.{
    ..color:.red;
    ..border:.1px.solid.black;
    ..margin:.10px;.}
    
    .box2.{
    ..color:.red;
    ..border:.1px.solid.black;
    ..margin:.10px;.}
    
    .box3.{
    ..color:.red;
    ..border:.1px.solid.black;
    ..margin:.10px;.}
    
    body.{
    ..font-family:."arial";.}
    
    /*#.sourceMappingURL=test-edit-insert-property.css.map.*/
    
===== test-edit-insert-property.scss =====
    $color:.red;
    
    @for.$i.from.1.through.3.{
    .....box#{$i}.{
    ........color:.$color;
    ........border:.1px.solid.black;
    ........margin:.10px;
    ....}
    }
    
    body.{
    ....font-family:."arial";
    }
    

---------------------- TEST: testInsertTrailingProperty ----------------------

Edits:
    {2, 10, 2, 15} '10px;' => '10px;border: 1px solid black;'
===== test-edit-insert-property.css =====
    .box1.{
    ..color:.red;
    ..margin:.10px;
    ..border:.1px.solid.black;.}
    
    .box2.{
    ..color:.red;
    ..margin:.10px;
    ..border:.1px.solid.black;.}
    
    .box3.{
    ..color:.red;
    ..margin:.10px;
    ..border:.1px.solid.black;.}
    
    body.{
    ..font-family:."arial";.}
    
    /*#.sourceMappingURL=test-edit-insert-property.css.map.*/
    
===== test-edit-insert-property.scss =====
    $color:.red;
    
    @for.$i.from.1.through.3.{
    .....box#{$i}.{
    ........color:.$color;
    ........margin:.10px;
    ........border:.1px.solid.black;
    ....}
    }
    
    body.{
    ....font-family:."arial";
    }
    

-------------------- TEST: testInsertDuplicateProperties ---------------------

Edits:
    {1, 2, 1, 7} 'color' => 'border: 1px solid black;color'
    {9, 2, 9, 7} 'color' => 'border: 1px solid black;color'
===== test-edit-insert-property.css =====
    .box1.{
    ..border:.1px.solid.black;
    ..color:.red;
    ..margin:.10px;.}
    
    .box2.{
    ..border:.1px.solid.black;
    ..color:.red;
    ..margin:.10px;.}
    
    .box3.{
    ..border:.1px.solid.black;
    ..color:.red;
    ..margin:.10px;.}
    
    body.{
    ..font-family:."arial";.}
    
    /*#.sourceMappingURL=test-edit-insert-property.css.map.*/
    
===== test-edit-insert-property.scss =====
    $color:.red;
    
    @for.$i.from.1.through.3.{
    .....box#{$i}.{
    ........border:.1px.solid.black;
    ........color:.$color;
    ........margin:.10px;
    ....}
    }
    
    body.{
    ....font-family:."arial";
    }
    

---------------- TEST: testPrependDifferentPropertiesSameRule ----------------

Edits:
    {1, 2, 1, 7} 'color' => 'display:flex;overflow:hidden;color'
===== test-edit-insert-property.css =====
    .box1.{
    ..display:.flex;
    ..overflow:.hidden;
    ..color:.red;
    ..margin:.10px;.}
    
    .box2.{
    ..display:.flex;
    ..overflow:.hidden;
    ..color:.red;
    ..margin:.10px;.}
    
    .box3.{
    ..display:.flex;
    ..overflow:.hidden;
    ..color:.red;
    ..margin:.10px;.}
    
    body.{
    ..font-family:."arial";.}
    
    /*#.sourceMappingURL=test-edit-insert-property.css.map.*/
    
===== test-edit-insert-property.scss =====
    $color:.red;
    
    @for.$i.from.1.through.3.{
    .....box#{$i}.{
    ........display:.flex;
    ........overflow:.hidden;
    ........color:.$color;
    ........margin:.10px;
    ....}
    }
    
    body.{
    ....font-family:."arial";
    }
    

------------ TEST: testInsertTrailingDifferentPropertiesSameRule -------------

Edits:
    {2, 10, 2, 15} '10px;' => '10px;display:flex;overflow:hidden;'
===== test-edit-insert-property.css =====
    .box1.{
    ..color:.red;
    ..margin:.10px;
    ..display:.flex;
    ..overflow:.hidden;.}
    
    .box2.{
    ..color:.red;
    ..margin:.10px;
    ..display:.flex;
    ..overflow:.hidden;.}
    
    .box3.{
    ..color:.red;
    ..margin:.10px;
    ..display:.flex;
    ..overflow:.hidden;.}
    
    body.{
    ..font-family:."arial";.}
    
    /*#.sourceMappingURL=test-edit-insert-property.css.map.*/
    
===== test-edit-insert-property.scss =====
    $color:.red;
    
    @for.$i.from.1.through.3.{
    .....box#{$i}.{
    ........color:.$color;
    ........margin:.10px;
    ........display:.flex;
    ........overflow:.hidden;
    ....}
    }
    
    body.{
    ....font-family:."arial";
    }
    

------------- TEST: testInsertDifferentPropertiesDifferentRules --------------

Edits:
    {1, 2, 1, 7} 'color' => 'margin:10px;color'
    {10, 10, 10, 15} '10px;' => '10px;border: 1px solid black;'
===== test-edit-insert-property.css =====
    .box1.{
    ..margin:.10px;
    ..color:.red;
    ..margin:.10px;
    ..border:.1px.solid.black;.}
    
    .box2.{
    ..margin:.10px;
    ..color:.red;
    ..margin:.10px;
    ..border:.1px.solid.black;.}
    
    .box3.{
    ..margin:.10px;
    ..color:.red;
    ..margin:.10px;
    ..border:.1px.solid.black;.}
    
    body.{
    ..font-family:."arial";.}
    
    /*#.sourceMappingURL=test-edit-insert-property.css.map.*/
    
===== test-edit-insert-property.scss =====
    $color:.red;
    
    @for.$i.from.1.through.3.{
    .....box#{$i}.{
    ........margin:.10px;
    ........color:.$color;
    ........margin:.10px;
    ........border:.1px.solid.black;
    ....}
    }
    
    body.{
    ....font-family:."arial";
    }
    
